requirements/optional.txt
celery>=3.1,<4.0
kombu>=3.0.37,<3.1
boto3
django-celery>=3.3.1
xlsxwriter>=1.3.6
factory-boy>=3.1.0
